requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
undetected-chromedriver>=3.5.5
beautifulsoup4>=4.12.0
Pillow>=10.0.0
tesserocr>=2.7.0
numpy>=1.26.0
pandas>=2.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
rich>=13.0.0
click>=8.1.0
tqdm>=4.66.0
fastapi>=0.110.0
uvicorn>=0.27.0
openai>=1.30.0
python-dotenv>=1.0.0
